Update JC_Button.cpp
The result of the digitalRead(button_pin) is only used in the else part of the if-else statement in the read() function. Therefore, the digitalRead and inversion check is moved to that else part of the if-else statement. By eliminating the time to do the digitalRead and inversion check during the debouncing period, the entire read() function is able to run faster on average. Was able to get a measurable difference through a test code.